Some changes in the functionality, the search

+++ BookRating/BookRating/DTOs/ModifiedReviewDto.cs
@@ -5,7 +5,7 @@ namespace BookRating.DTOs
     public class ModifiedReviewDto
     {
         [Range(1, 5)]
-        public double? Rating { get; set; }
+        public int? Rating { get; set; }
         public string? Comment { get; set; }
     }
 }
